Fix and enhance NAT checksum update (#397)
* nat: Fix icmpv4 checksum.

The icmpv4 header checksum shouldn't include the pseudoheader.

Type and code are 8-bit, not 16.

* nat: Use incremental checksum updates.

* nat: Rewrite only what's required.

* utils/checksum: Fix comment

* nat: Add basic module test.

+++ core/utils/icmp.h
@@ -5,8 +5,8 @@ namespace utils {
 
 // A basic ICMP header definition.
 struct[[gnu::packed]] IcmpHeader {
-  uint16_t type;      // ICMP packet type.
-  uint16_t code;      // ICMP packet type.
+  uint8_t type;       // ICMP packet type.
+  uint8_t code;       // ICMP packet code.
   uint16_t checksum;  // ICMP packet checksum.
   uint16_t ident;     // ICMP packet identifier.
   uint16_t seq_num;   // ICMP packet sequence number
